arduino_driver: make serial interface a separate module
